RCA-13 Removes stderr messages from multiprocessing.
Lowers "triggering flush" level to warning.

The stderr messages are mostly clutter, and I don't believe provide
valuable enough information for the user.

The flush trigger does not need to be critical, as a critical indicates
something is horribly wrong with the the application. A warning should
hopefully suffice, letting the user know that a flush is happening but it is
expected behavior since the link is down. This is something we can discuss.